+++ addons/hr/models/__init__.py
@@ -10,6 +10,7 @@ from . import hr_job
 from . import hr_plan
 from . import mail_alias
 from . import mail_channel
+from . import models
 from . import res_config_settings
 from . import res_partner
 from . import res_users
