+++ addons/hr/models/mail_alias.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import fields, models, tools, _
+from odoo import fields, models, _
 
 
 class Alias(models.Model):
@@ -16,21 +16,4 @@ class Alias(models.Model):
             return _("""Hi,<br/>
 Your document has not been created because your email address is not recognized.<br/>
 Please send emails with the email address recorded on your employee information, or contact your HR manager.""")
-        else:
-            return super(Alias, self)._get_alias_bounced_body_fallback(message_dict)
-
-
-class AliasMixin(models.AbstractModel):
-    _inherit = 'mail.alias.mixin'
-
-    def _alias_check_contact_on_record(self, record, message, message_dict, alias):
-        if alias.alias_contact == 'employees':
-            email_from = tools.decode_message_header(message, 'From')
-            email_address = tools.email_split(email_from)[0]
-            employee = self.env['hr.employee'].search([('work_email', 'ilike', email_address)], limit=1)
-            if not employee:
-                employee = self.env['hr.employee'].search([('user_id.email', 'ilike', email_address)], limit=1)
-            if not employee:
-                return _('restricted to employees')
-            return True
-        return super(AliasMixin, self)._alias_check_contact_on_record(record, message, message_dict, alias)
+        return super(Alias, self)._get_alias_bounced_body_fallback(message_dict)
