+++ addons/mail/wizard/mail_compose_message.py
@@ -278,7 +278,7 @@ class MailComposer(models.TransientModel):
         reply_to_value = dict.fromkeys(res_ids, None)
         if mass_mail_mode and not self.no_auto_thread:
             records = self.env[self.model].browse(res_ids)
-            reply_to_value = self.env['mail.thread']._notify_get_reply_to_on_records(default=self.email_from, records=records)
+            reply_to_value = records._notify_get_reply_to(default=self.email_from)
 
         blacklisted_rec_ids = []
         if mass_mail_mode and issubclass(type(self.env[self.model]), self.pool['mail.thread.blacklist']):
@@ -462,7 +462,7 @@ class MailComposer(models.TransientModel):
         default_recipients = {}
         if not self.partner_ids:
             records = self.env[self.model].browse(res_ids).sudo()
-            default_recipients = self.env['mail.thread']._message_get_default_recipients_on_records(records)
+            default_recipients = records._message_get_default_recipients()
 
         results = dict.fromkeys(res_ids, False)
         for res_id in res_ids:
